Add completed tests for JobDeleteCommand.java
